# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sabor & Gestión API
  description: API REST del sistema Sabor & Gestión para la gestión operativa del restaurante.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local
...
paths:
...
  /api/pedidos:
    get:
      tags:
        - Pedidos
      summary: Listar pedidos con filtros
      description: 'Obtiene los pedidos del restaurante. Admite filtrado por hoy, fecha, mesa, activo, cajero, mesero, o reportesCierre=true para arqueos de 48h. Requiere autenticación (personal operativo: Mesero, Cajero, Cocinero, Administrador).'
      operationId: obtenerPedidos
      security:
        - bearerAuth: []
      parameters:
        - name: hoy
          in: query
          required: false
          description: Filtrar pedidos del día de hoy
          schema:
            type: boolean
        - name: fecha
          in: query
          required: false
          description: Filtrar por fecha específica (YYYY-MM-DD)
          schema:
            type: string
            format: date
          example: '2026-09-27'
        - name: mesa
          in: query
          required: false
          description: ID de la mesa para filtrar sus pedidos
          schema:
            type: string
          example: '{{mesaId}}'
        - name: activo
          in: query
          required: false
          description: Filtrar solo pedidos activos (ABIERTO, EN_PREPARACION, ENTREGADO, CUENTA_SOLICITADA)
          schema:
            type: boolean
        - name: cajero
          in: query
          required: false
          description: ID del cajero asignado
          schema:
            type: string
          example: '{{usuarioId}}'
        - name: mesero
          in: query
          required: false
          description: ID del usuario mesero que creó el pedido
          schema:
            type: string
          example: '{{usuarioId}}'
        - name: reportesCierre
          in: query
          required: false
          description: Retorna los reportes de cierre de caja registrados en las últimas 48 horas
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de pedidos o reportes de cierre
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pedido'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al obtener pedidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
    post:
      tags:
        - Pedidos
      summary: Crear un nuevo pedido
      description: Crea un pedido validando previamente la disponibilidad de los ingredientes requeridos según las recetas. El backend calcula los subtotales e importes con los precios oficiales de Plato obtenidos vía PlatoRepository y almacena el snapshot histórico de precioUnitario; los importes enviados por el cliente no son confiables y se recalculan. Si se envía mesa y no existe en la base de datos, retorna 404. Genera el código correlativo PED-XXXX y cambia automáticamente el estado de la mesa a "Ocupada". Requiere rol Mesero o Administrador.
      operationId: crearPedido
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearPedidoRequest'
      responses:
        '201':
          description: Pedido registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '400':
          description: Error de validación (ingredientes no disponibles, montos negativos o cantidades inválidas)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoErrorIngredientesResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere rol Mesero o Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Mesa no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al registrar pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
  /api/pedidos/pendientes-cobro:
    get:
      tags:
        - Pedidos
      summary: Listar pedidos pendientes de cobro para Caja
      description: Obtiene las órdenes activas cuyas mesas están en estado "Cuenta Solicitada", estructuradas con subtotales, propinas y desglose de ítems para el cobro. Requiere rol Cajero o Administrador.
      operationId: obtenerPedidosPendientesCobro
      security:
        - bearerAuth: []
      parameters:
        - name: cajero
          in: query
          required: false
          description: ID de cajero asignado
          schema:
            type: string
          example: '{{usuarioId}}'
      responses:
        '200':
          description: Lista de pedidos en espera de cobro
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PayloadCajaDTO'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere rol Cajero o Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al obtener pedidos pendientes de cobro
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
  /api/pedidos/{id}:
    put:
      tags:
        - Pedidos
      summary: Actualizar contenido de un pedido
      description: Actualiza platos (detalles), descuentos, propinas o datos del comensal. El backend recalcula los subtotales e importes con los precios oficiales de Plato obtenidos vía PlatoRepository y almacena el snapshot histórico; no confía en montos provistos por el cliente. Si el pedido estaba en ENTREGADO y se añaden platos, se reabre a ABIERTO y la mesa se reactiva como Ocupada. Requiere rol Mesero o Administrador.
      operationId: actualizarPedido
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador del pedido
          schema:
            type: string
          example: '{{pedidoId}}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarPedidoRequest'
      responses:
        '200':
          description: Pedido actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '400':
          description: Error de validación (montos negativos o cantidades inválidas)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere rol Mesero o Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al actualizar el pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
  /api/pedidos/{id}/estado:
    patch:
      tags:
        - Pedidos
      summary: Actualizar estado del pedido (Cocina)
      description: 'Actualiza el estado operativo del pedido en cocina: ABIERTO -> EN_PREPARACION -> ENTREGADO. Al pasar a ENTREGADO dispara la alerta "¡Listo!" a meseros por WebSocket. Requiere rol Cocinero o Administrador.'
      operationId: actualizarEstadoPedido
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador del pedido
          schema:
            type: string
          example: '{{pedidoId}}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarEstadoPedidoRequest'
      responses:
        '200':
          description: Estado del pedido actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActualizarEstadoPedidoResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere rol Cocinero o Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al actualizar estado del pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
  /api/pedidos/{id}/solicitar-cuenta:
    patch:
      tags:
        - Pedidos
      summary: Solicitar cuenta de un pedido
      description: Solicita la cuenta de la orden, cambiando el estado de la mesa a "Cuenta Solicitada" (notificando estado de WebSocket "Esperando pago" a los clientes) y notificando a caja en tiempo real. Requiere rol Mesero o Administrador.
      operationId: solicitarCuentaPedido
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador del pedido
          schema:
            type: string
          example: '{{pedidoId}}'
      responses:
        '200':
          description: Cuenta solicitada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitarCuentaResponse'
        '400':
          description: Pedido sin mesa o pedido ya cerrado/cancelado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere rol Mesero o Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Pedido o mesa no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al solicitar la cuenta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
  /api/pedidos/{id}/cancel:
    patch:
      tags:
        - Pedidos
      summary: Cancelar pedido y liberar mesa
      description: Cancela un pedido cambiando su estado a CANCELADO y libera la mesa poniéndola en "Libre" (emitiendo statusSocket "Disponible" en WebSocket). Requiere rol Mesero o Administrador.
      operationId: cancelarPedido
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador del pedido
          schema:
            type: string
          example: '{{pedidoId}}'
      responses:
        '200':
          description: Pedido cancelado y mesa liberada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelarPedidoResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere rol Mesero o Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al cancelar pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Autenticación mediante token JWT en cabecera: Authorization: Bearer <token>'
  schemas:
    MensajeErrorResponse:
      type: object
      required:
        - mensaje
      properties:
        mensaje:
          type: string
...
    UsuarioNormalizado:
      type: object
      properties:
        id:
          type: string
          example: '000000000000000000000001'
        _id:
          type: string
          example: '000000000000000000000001'
        nombre:
          type: string
          example: Juan
        apellido:
          type: string
          example: Pérez
        ci:
          type: string
          example: '12345678'
        email:
          type: string
          format: email
          example: juan.perez@example.com
        rol:
          type: string
          enum:
            - Administrador
            - Mesero
            - Cocinero
            - Cajero
          example: Mesero
        estado:
          type: boolean
          example: true
        verificado:
          type: boolean
          example: true
        zona:
          type: string
          example: Salón Principal
        ubicacion:
          type: string
          example: Salón Principal
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MesaDTO:
      type: object
      properties:
        id:
          type: string
          example: '000000000000000000000030'
        _id:
          type: string
          example: '000000000000000000000030'
        name:
          type: string
          example: Mesa 1 Interior
        numero:
          type: string
          example: Mesa 1 Interior
        capacity:
          type: integer
          example: 4
        location:
          type: string
          example: Interior
        locationId:
          type: string
          nullable: true
          example: '000000000000000000000040'
        status:
          type: string
          example: Disponible
          description: Mapeo para visualización en frontend y eventos WebSocket (Disponible, Ocupada, Esperando pago)
        type:
          type: string
          enum:
            - normal
            - vip
          example: normal
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    DetallePedido:
      type: object
      required:
        - plato
        - cantidad
        - precioUnitario
        - subtotal
      properties:
        plato:
          oneOf:
            - type: string
              example: '000000000000000000000020'
            - $ref: '#/components/schemas/Plato'
        nombrePlato:
          type: string
          example: Lomo Montado
        cantidad:
          type: integer
          minimum: 1
          example: 2
        precioUnitario:
          type: number
          minimum: 0
          example: 45
          description: Snapshot histórico del precio unitario del plato obtenido por el backend al momento del pedido
        subtotal:
          type: number
          minimum: 0
          example: 90
          description: Subtotal de la línea (precioUnitario × cantidad)
        observacion:
          type: string
          example: Término medio
    CrearDetallePedidoInput:
      type: object
      required:
        - plato
        - cantidad
      properties:
        plato:
          type: string
          example: '{{platoId}}'
          description: ID del plato a ordenar
        cantidad:
          type: integer
          minimum: 1
          example: 2
          description: Cantidad de porciones (entero >= 1)
        precioUnitario:
          type: number
          minimum: 0
          example: 45
          description: Opcional / Informativo. El backend no confía en este valor; consulta el precio oficial de Plato en BD vía PlatoRepository y almacena el snapshot histórico.
        subtotal:
          type: number
          minimum: 0
          example: 90
          description: Opcional / Informativo. El backend recalcula subtotal = precioUnitario × cantidad.
        observacion:
          type: string
          example: Sin cebolla
          description: Nota o exclusión de ingredientes (ej. "sin Cebolla")
    Pedido:
      type: object
      properties:
        _id:
          type: string
          example: '000000000000000000000070'
        codigo:
          type: string
          example: PED-0001
        fechaDiaBolivia:
          type: string
          example: '2026-09-27'
        fechaHoraBolivia:
          type: string
          example: 27/09/2026 14:30:00
        fechaHora:
          type: string
          format: date-time
        estado:
          type: string
          enum:
            - ABIERTO
            - EN_PREPARACION
            - ENTREGADO
            - CANCELADO
            - CERRADO
          example: ABIERTO
        total:
          type: number
          minimum: 0
          example: 90
        mesa:
          oneOf:
            - type: string
              example: '000000000000000000000030'
            - $ref: '#/components/schemas/MesaDTO'
        usuario:
          oneOf:
            - type: string
              example: '000000000000000000000001'
            - $ref: '#/components/schemas/UsuarioNormalizado'
        detalles:
          type: array
          items:
            $ref: '#/components/schemas/DetallePedido'
        qrUrl:
          type: string
          example: https://api.qrserver.com/v1/create-qr-code/?...
        metodoPago:
          type: string
          enum:
            - Efectivo
            - Tarjeta
            - Transferencia
            - QR
            - Otro
          example: Efectivo
        montoDescuento:
          type: number
          minimum: 0
          example: 0
        montoPropina:
          type: number
          minimum: 0
          example: 0
        subtotalCierre:
          type: number
          minimum: 0
          example: 90
        clienteNombre:
          type: string
          example: Carlos Mendoza
        clienteCI:
          type: string
          example: '12345678'
        clienteNIT:
          type: string
          example: '12345678019'
        cajeroAsignado:
          type: string
          example: '000000000000000000000001'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CrearPedidoRequest:
      type: object
      required:
        - detalles
      properties:
        mesa:
          type: string
          example: '{{mesaId}}'
          description: ID de la mesa asignada (si se especifica y no existe en BD, retorna 404)
        detalles:
          type: array
          items:
            $ref: '#/components/schemas/CrearDetallePedidoInput'
          description: Lista de platos y cantidades a ordenar
        montoDescuento:
          type: number
          minimum: 0
          example: 0
          description: Monto de descuento a aplicar (no puede ser negativo)
        montoPropina:
          type: number
          minimum: 0
          example: 0
          description: Monto de propina voluntaria (no puede ser negativo)
        total:
          type: number
          minimum: 0
          example: 90
          description: Opcional / Informativo. El backend calcula el total automáticamente aplicando la fórmula Math.max(0, subtotal - montoDescuento + montoPropina).
        clienteNombre:
          type: string
          example: Consumidor Final
          description: Nombre del comensal o cliente en mesa
        clienteCI:
          type: string
          example: '12345678'
          description: Cédula de identidad del cliente
        clienteNIT:
          type: string
          example: '12345678019'
          description: NIT del cliente para facturación
    PedidoErrorIngredientesResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        mensaje:
          type: string
          example: No se puede registrar el pedido por restricciones de receta o ingredientes
        errores:
          type: array
          items:
            type: string
          example:
            - 'El plato "Lomo Montado" no se puede preparar porque falta: Lomo de Res'
        faltantes:
          type: array
          items:
            type: string
          example:
            - Lomo de Res
    ActualizarPedidoRequest:
      type: object
      properties:
        detalles:
          type: array
          items:
            $ref: '#/components/schemas/CrearDetallePedidoInput'
          description: Nuevos detalles del pedido. Si se proveen, el backend recalcula subtotales, totales y snapshot con precios oficiales de Plato en BD.
        estado:
          type: string
          enum:
            - ABIERTO
            - EN_PREPARACION
            - ENTREGADO
            - CANCELADO
            - CERRADO
          example: ABIERTO
        metodoPago:
          type: string
          enum:
            - Efectivo
            - Tarjeta
            - Transferencia
            - QR
            - Otro
          example: Efectivo
        montoDescuento:
          type: number
          minimum: 0
          example: 10
          description: Monto de descuento (no puede ser negativo)
        montoPropina:
          type: number
          minimum: 0
          example: 5
          description: Monto de propina (no puede ser negativo)
        subtotalCierre:
          type: number
          minimum: 0
          example: 140
          description: Subtotal de cierre (no puede ser negativo)
        total:
          type: number
          minimum: 0
          example: 135
          description: Opcional / Informativo. El backend recalcula el total automáticamente si se modifican detalles, descuento o propina.
        clienteNombre:
          type: string
          example: Carlos Mendoza
        clienteCI:
          type: string
          example: '87654321'
        clienteNIT:
          type: string
          example: '87654321012'
        cajeroAsignado:
          type: string
          example: '{{usuarioId}}'
    ActualizarEstadoPedidoRequest:
      type: object
      required:
        - estado
      properties:
        estado:
          type: string
          enum:
            - ABIERTO
            - EN_PREPARACION
            - ENTREGADO
            - CERRADO
            - CANCELADO
          example: EN_PREPARACION
    ActualizarEstadoPedidoResponse:
      type: object
      properties:
        mensaje:
          type: string
          example: Pedido movido a EN_PREPARACION
        pedido:
          $ref: '#/components/schemas/Pedido'
    CancelarPedidoResponse:
      type: object
      properties:
        mensaje:
          type: string
          example: Pedido anulado y mesa liberada correctamente
        pedido:
          $ref: '#/components/schemas/Pedido'
    ItemPayloadCaja:
      type: object
      properties:
        platoId:
          type: string
          example: '000000000000000000000020'
        nombre:
          type: string
          example: Lomo Montado
        cantidad:
          type: integer
          example: 2
        precioUnitario:
          type: number
          example: 45
        subtotal:
          type: number
          example: 90
        observacion:
          type: string
          example: ''
    PayloadCajaDTO:
      type: object
      properties:
        pedidoId:
          type: string
          example: '000000000000000000000070'
        codigo:
          type: string
          example: PED-0001
        mesa:
          type: string
          example: Mesa 1 Interior
        mesero:
          type: string
          example: Juan Pérez
        estado:
          type: string
          example: ABIERTO
        subtotal:
          type: number
          example: 90
        itemsSubtotal:
          type: number
          example: 90
        descuento:
          type: number
          example: 0
        montoDescuento:
          type: number
          example: 0
        propina:
          type: number
          example: 0
        montoPropina:
          type: number
          example: 0
        total:
          type: number
          example: 90
        clienteNombre:
          type: string
          example: Consumidor Final
        clienteCI:
          type: string
          example: ''
        clienteNIT:
          type: string
          example: ''
        tiempoEsperaMinutos:
          type: integer
          example: 15
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemPayloadCaja'
        fechaHoraBolivia:
          type: string
          example: 27/09/2026 14:30:00
    SolicitarCuentaResponse:
      type: object
      properties:
        mensaje:
          type: string
          example: Cuenta solicitada correctamente
        solicitud:
          $ref: '#/components/schemas/PayloadCajaDTO'
...